% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ceg.R
\name{compare_ceg_models}
\alias{compare_ceg_models}
\title{Compare Two Chain Event Graph Models}
\usage{
compare_ceg_models(ceg1, ceg2)
}
\arguments{
\item{ceg1}{An object of class
\code{"ceg"}.}

\item{ceg2}{An object of class
\code{"ceg"}.}
}
\value{
An object of class \code{"compare_ceg_models"} containing:
\itemize{
\item Log marginal likelihoods for both models.
\item Log Bayes Factor.
\item Bayes Factor.
\item Preferred model.
\item Evidence measures for each model.
\item Jeffreys evidence category.
}
}
\description{
Compares two fitted Chain Event Graph (CEG) models using their log marginal
likelihoods and calculates the corresponding Bayes Factor.
}
\details{
Evidence in favour of each model is quantified and classified according to
Jeffreys' scale for Bayes Factors.

The Bayes Factor is computed as:

\deqn{
BF = \exp(\log p(D \mid M_1) - \log p(D \mid M_2))
}

where \eqn{p(D \mid M)} denotes the marginal likelihood of a model.

The resulting Bayes Factor is interpreted using Jeffreys' evidence scale.
}
\examples{
\dontrun{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg_model1 <- compute_ceg(st_priors)

priors2 <- specify_priors(
  st,
  prior_type = "Phantom"
)

st_priors2 <- compute_staged_tree_priors(
  st,
  priors2
)

ceg_model2 <- compute_ceg(st_priors2)

comparison <- compare_ceg_models(
  ceg_model1,
  ceg_model2
)

print(comparison)
summary(comparison)
}

}
\seealso{
\code{\link{summary.compare_ceg_models}}
}
