% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_priors.R
\name{edit_priors}
\alias{edit_priors}
\title{Edit Priors in a Prior Table}
\usage{
edit_priors(prior_table, rows, new_priors)
}
\arguments{
\item{prior_table}{An object of class \code{"prior_table"}.}

\item{rows}{Integer vector specifying rows to modify.}

\item{new_priors}{List containing replacement prior specifications.}
}
\value{
An updated object of class \code{"prior_table"}.
}
\description{
Modifies one or more rows of a prior table created by
\code{\link{specify_priors}}.
}
\details{
The function validates that the number of supplied Dirichlet parameters
matches the number of outgoing edges associated with each stage.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)
priors <- specify_priors(st, "Uniform")

priors <- edit_priors(
  priors,
  rows = 1,
  new_priors = list("2,3,4,5")
)

print(priors)

}
\seealso{
\code{\link{specify_priors}}
}
