% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homicides.R
\docType{data}
\name{homicides}
\alias{homicides}
\title{Homicides Dataset}
\format{
A data frame with 2670 rows and 9 variables:
\describe{
\item{Age_Group}{Age group of the victim. One of: \code{"Adult"}, \code{"Adolescent/Young Adult"}, \code{"Child"}, \code{"Elderly"}.}
\item{Sex}{Sex of the victim. \code{"Male"} or \code{"Female"}}
\item{Method_of_Killing}{Recorded method of killing, e.g., \code{"Knife or Sharp Implement"}, \code{"Blunt Implement"}, etc.}
\item{Domestic_Abuse}{Indicates whether the incident was flagged as domestic abuse.}
\item{Solved_Status}{Whether the case has been solved. One of: \code{"Solved"}, \code{"Unsolved"}}
\item{Borough}{London Borough where the homicide occurred}
\item{Ethnicity}{Recorded ethnicity of the victim. One of \code{"Asian"}, \code{"Black"}, \code{"Not Reported/Not Known"}, \code{"Other"}, \code{"White"}.}
\item{Year}{Year in which the incident was recorded.}
\item{BCU}{Basic Command Unit where the homicide occurred (Homicide units changed from Boroughs to BCUs in 2017).}
}
}
\source{
Data taken and filtered from https://www.met.police.uk/police-forces/metropolitan-police/areas/stats-and-data/stats-and-data/met/homicide-dashboard/
}
\usage{
homicides
}
\description{
A dataset containing homicides recorded by the Metropolitan Police from 2003-2023
}
\keyword{datasets}
