% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotting.R
\name{plot.ceg_map}
\alias{plot.ceg_map}
\title{Plot a CEG Probability Map}
\usage{
\method{plot}{ceg_map}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ceg_map"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
A leaflet map widget.
}
\description{
Displays the interactive leaflet map stored within a
\code{"ceg_map"} object.
}
\examples{
et <- create_event_tree(homicides, c(9,1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

map_obj <- generate_CEG_map(
  shapefile = bcu_shapefile,
  ceg_object = ceg,
  colour_by = "Female"
)

plot(map_obj)


}
