% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_tree.R
\name{plot.event_tree}
\alias{plot.event_tree}
\title{Plot an Event Tree}
\usage{
\method{plot}{event_tree}(x, label_type = "both", level_separation = 1600, node_distance = 400, ...)
}
\arguments{
\item{x}{An object of class \code{"event_tree"}.}

\item{label_type}{Character string specifying the edge label format.
Supported values are:
\itemize{
\item \code{"both"} (default), displaying labels and frequencies.
\item \code{"names"}, displaying labels only.
}}

\item{level_separation}{Numeric value controlling spacing between levels.
Default is \code{1600}.}

\item{node_distance}{Numeric value controlling spacing between nodes.
Default is \code{400}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
A \pkg{visNetwork} htmlwidget.
}
\description{
Produces an interactive visualisation of an event tree using
\pkg{visNetwork}.
}
\details{
Edge labels may display either transition names only or both transition
names and observed frequencies.
}
\examples{
\dontrun{
et <- create_event_tree(homicides, c(1:3))

plot(et)
}

}
\seealso{
\code{\link{create_event_tree}}
}
