% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged_tree_prior.R
\name{plot.staged_tree_priors}
\alias{plot.staged_tree_priors}
\title{Plot a Staged Tree with Priors}
\usage{
\method{plot}{staged_tree_priors}(
  x,
  level_separation = 1600,
  node_distance = 400,
  label_type = "priors",
  ...
)
}
\arguments{
\item{x}{An object of class \code{"staged_tree_priors"}.}

\item{level_separation}{Numeric value controlling spacing between graph
levels. Default is \code{1600}.}

\item{node_distance}{Numeric value controlling spacing between nodes.
Default is \code{400}.}

\item{label_type}{Character string specifying the edge labels to display.
Supported values are:
\itemize{
\item \code{"priors"} (default)
\item \code{"priormeans"}
\item \code{"names"}
}}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
A \pkg{visNetwork} htmlwidget.
}
\description{
Produces an interactive visualisation of a staged tree containing prior
information using \pkg{visNetwork}.
}
\details{
Node tooltips display prior distributions, prior means and prior variances.
Edge labels may display prior values or prior means.
}
\examples{

et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

plot(st_priors)


}
\seealso{
\code{\link{compute_staged_tree_priors}},
\code{\link{compute_ceg}}
}
