% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ceg.R
\name{print.ceg}
\alias{print.ceg}
\title{Print a Chain Event Graph}
\usage{
\method{print}{ceg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ceg"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied \code{"ceg"} object, invisibly.
}
\description{
Prints a concise overview of a Chain Event Graph including the number of
vertices, edges and stages.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

print(ceg)

}
