% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotting.R
\name{print.ceg_map}
\alias{print.ceg_map}
\title{Print a CEG Probability Map}
\usage{
\method{print}{ceg_map}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ceg_map"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied \code{"ceg_map"} object, invisibly.
}
\description{
Prints a concise description of a \code{"ceg_map"} object including the
outcome being mapped, probability range and the number of excluded
geographical regions.
}
\examples{
\dontrun{
et <- create_event_tree(homicides, c(9,1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

map_obj <- generate_CEG_map(
  shapefile = bcu_shapefile,
  ceg_object = ceg,
  colour_by = "Female"
)

print(map_obj)
}

}
