% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_tree.R
\name{print.event_tree}
\alias{print.event_tree}
\title{Print an Event Tree}
\usage{
\method{print}{event_tree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"event_tree"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied \code{"event_tree"} object, invisibly.
}
\description{
Prints a concise summary of an event tree including the variables used,
number of nodes and number of edges.
}
\examples{
et <- create_event_tree(homicides, c(1:3))

print(et)

}
