% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_priors.R
\name{print.prior_table}
\alias{print.prior_table}
\title{Print a Prior Table}
\usage{
\method{print}{prior_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"prior_table"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied object, invisibly.
}
\description{
Prints the stage-level prior table contained within a
\code{"prior_table"} object.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)
priors <- specify_priors(st, "Uniform")

print(priors)

}
