% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reduced_CEG.R
\name{print.reduced_ceg}
\alias{print.reduced_ceg}
\title{Print a Reduced Chain Event Graph}
\usage{
\method{print}{reduced_ceg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"reduced_ceg"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied \code{"reduced_ceg"} object, invisibly.
}
\description{
Prints a concise summary of a reduced Chain Event Graph including the number
of nodes, edges and starting labels used to construct the graph.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

reduced_ceg <- compute_reduced_ceg(
  ceg,
  start_labels = "Adult"
)

print(reduced_ceg)


}
