% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_staged_tree.R
\name{print.staged_tree}
\alias{print.staged_tree}
\title{Print a Staged Tree}
\usage{
\method{print}{staged_tree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"staged_tree"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied \code{"staged_tree"} object, invisibly.
}
\description{
Prints a concise summary of a staged tree, including the number of nodes,
edges and colouring methods used.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

print(st)

}
