% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged_tree_prior.R
\name{print.staged_tree_priors}
\alias{print.staged_tree_priors}
\title{Print a Staged Tree with Priors}
\usage{
\method{print}{staged_tree_priors}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"staged_tree_priors"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
The supplied object, invisibly.
}
\description{
Prints a concise overview of a \code{"staged_tree_priors"} object,
including the number of nodes, edges and prior type.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

print(st_priors)

}
