% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{stCEG-package}
\alias{stCEG}
\alias{stCEG-package}
\title{stCEG}
\description{
Tools for constructing Event Trees, Staged Trees and
Chain Event Graphs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/holliecalley/stCEG}
  \item Report bugs at \url{https://github.com/holliecalley/stCEG/issues}
}

}
\author{
\strong{Maintainer}: Hollie Calley \email{hc629@exeter.ac.uk} (\href{https://orcid.org/0009-0002-1814-8428}{ORCID})

Authors:
\itemize{
  \item Hollie Calley \email{hc629@exeter.ac.uk} (\href{https://orcid.org/0009-0002-1814-8428}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel Williamson (\href{https://orcid.org/0000-0001-8917-3300}{ORCID}) [contributor]
}

}
\keyword{internal}
