% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ceg.R
\name{summary.ceg}
\alias{summary.ceg}
\title{Summarise a Chain Event Graph}
\usage{
\method{summary}{ceg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ceg"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
An object of class \code{"summary_ceg"} containing:
\itemize{
\item Number of vertices.
\item Number of edges.
\item Graph levels.
\item Stage colours.
\item Stage identifiers.
\item Prior type information.
}
}
\description{
Generates a concise summary of a Chain Event Graph including the number of
contracted vertices, edges, stages, levels and available prior and posterior
information.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

summary(ceg)

}
