% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plotting.R
\name{summary.ceg_map}
\alias{summary.ceg_map}
\title{Summarise a CEG Probability Map}
\usage{
\method{summary}{ceg_map}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ceg_map"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
An object of class \code{"summary_ceg_map"}.
}
\description{
Produces a summary of a \code{"ceg_map"} object, including the number of
mapped areas, probability range and any excluded polygons.
}
\examples{
\dontrun{
et <- create_event_tree(homicides, c(9,1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

map_obj <- generate_CEG_map(
  shapefile = bcu_shapefile,
  ceg_object = ceg,
  colour_by = "Female"
)

summary(map_obj)
}

}
