% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ceg.R
\name{summary.compare_ceg_models}
\alias{summary.compare_ceg_models}
\title{Summarise a CEG Model Comparison}
\usage{
\method{summary}{compare_ceg_models}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"compare_ceg_models"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
An object of class \code{"summary_compare_ceg_models"} containing Bayes
Factor statistics, preferred model information and evidence measures.
}
\description{
Produces a summary of a Bayes Factor comparison between two Chain Event
Graph models.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg_model1 <- compute_ceg(st_priors)

priors2 <- specify_priors(
  st,
  prior_type = "Phantom"
)

st_priors2 <- compute_staged_tree_priors(
  st,
  priors2
)

ceg_model2 <- compute_ceg(st_priors2)

comparison <- compare_ceg_models(
  ceg_model1,
  ceg_model2
)

summary(comparison)

}
