% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_tree.R
\name{summary.event_tree}
\alias{summary.event_tree}
\title{Summarise an Event Tree}
\usage{
\method{summary}{event_tree}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"event_tree"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
A list containing:
\itemize{
\item Variables used to construct the tree.
\item Number of nodes.
\item Number of edges.
\item Preview of node information.
\item Preview of edge information.
}
}
\description{
Produces a summary of an event tree including variable names, graph size,
and previews of node and edge information.
}
\examples{
et <- create_event_tree(homicides, c(1:3))

summary(et)

}
