% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_priors.R
\name{summary.prior_table}
\alias{summary.prior_table}
\title{Summarise a Prior Table}
\usage{
\method{summary}{prior_table}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"prior_table"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
An object of class \code{"summary_prior_table"}.
}
\description{
Produces a summary of a prior table including the number of stages,
levels, colours and prior types represented.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)
priors <- specify_priors(st, "Uniform")

summary(priors)

}
