% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reduced_CEG.R
\name{summary.reduced_ceg}
\alias{summary.reduced_ceg}
\title{Summarise a Reduced Chain Event Graph}
\usage{
\method{summary}{reduced_ceg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"reduced_ceg"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
An object of class \code{"summary_reduced_ceg"} containing:
\itemize{
\item Number of nodes.
\item Number of edges.
\item Starting labels used for extraction.
\item Levels represented in the reduced graph.
\item Stage colours present in the reduced graph.
}
}
\description{
Produces a summary of a reduced Chain Event Graph including graph size,
levels present and stage colours represented in the extracted subgraph.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

ceg <- compute_ceg(st_priors)

reduced_ceg <- compute_reduced_ceg(
  ceg,
  start_labels = "Adult"
)
summary(reduced_ceg)


}
