% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_staged_tree.R
\name{summary.staged_tree}
\alias{summary.staged_tree}
\title{Summarise a Staged Tree}
\usage{
\method{summary}{staged_tree}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"staged_tree"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
A list containing:
\itemize{
\item Number of nodes.
\item Number of edges.
\item Preview of node information.
\item Preview of edge information.
\item Colouring methods used.
\item Primary colouring method.
}
}
\description{
Produces a summary of a staged tree, including graph size, colouring
methods and previews of node and edge information.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

summary(st)

}
