% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staged_tree_prior.R
\name{summary.staged_tree_priors}
\alias{summary.staged_tree_priors}
\title{Summarise a Staged Tree with Priors}
\usage{
\method{summary}{staged_tree_priors}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"staged_tree_priors"}.}

\item{...}{Additional arguments passed to S3 methods.}
}
\value{
An object of class \code{"summary_staged_tree_priors"} containing:
\itemize{
\item Number of nodes.
\item Number of edges.
\item Prior type information.
\item Numbers of uniform and custom priors.
\item Levels represented in the tree.
\item Stage colours.
}
}
\description{
Produces a summary of a \code{"staged_tree_priors"} object including graph
size, prior types and stage information.
}
\examples{
et <- create_event_tree(homicides, c(1:3))
st <- ahc_colouring(et)

priors <- specify_priors(
  st,
  prior_type = "Uniform"
)

st_priors <- compute_staged_tree_priors(
  st,
  priors
)

summary(st_priors)

}
