% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stceg.R
\name{run_stceg}
\alias{run_stceg}
\title{Run the stCEG Shiny App}
\usage{
run_stceg()
}
\value{
The function launches a Shiny app for spatial data analysis with interactive
visualizations and data manipulation features.
}
\description{
The app facilitates the upload and processing of spatial data, including the
dynamic selection of area divisions, time divisions, and prediction variables.
It integrates the leaflet package for interactive maps and visNetwork for visualizing
event trees. Additionally, the app supports the use of agglomerative hierarchical clustering (AHC)
Run the stCEG Shiny App

This function launches a Shiny app that allows users to upload spatial data
(in the form of a shapefile) and visualize it using Chain Event Graphs (CEGs).
The app enables the user to apply various spatial and temporal analyses,
including map visualizations and event tree colourings.

The app facilitates the upload and processing of spatial data, including the
dynamic selection of area divisions, time divisions, and prediction variables.
It integrates the leaflet package for interactive maps and visNetwork for visualizing
event trees. Additionally, the app supports the use of agglomerative hierarchical clustering (AHC)
for node colouring in event trees and allows for specification of prior distributions.
}
\details{
This function launches a Shiny app that allows users to upload spatial data
(in the form of a shapefile) and visualize it using Chain Event Graphs (CEGs).
The app enables the user to apply various spatial and temporal analyses,
including map visualizations and event tree colourings.
}
\examples{
if(interactive()){
# Run the app
run_stceg()}

}
